+++ packages/f-registration/CHANGELOG.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.35.0
 ------------------------------
-*October 9, 2020*
+*October 15, 2020*
 ### Added
 - Moved text into translations file 
 - Support for locale en-GB